tracking: move multi-camera C++ helpers into utils
