+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/config.cpp
@@ -1,4 +1,4 @@
-#include "examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/config_api.cpp"
+#include "examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/config_api.cpp"
 
 #include <algorithm>
 #include <cctype>
