+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/config_api.cpp
