+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/pipeline_api.cpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/config_api.cpp"
+#include "examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/config_api.cpp"
 #include "support/runtime/example_utils.h"
 
 #include "neat/models.h"
