+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/tracker.cpp
@@ -1,4 +1,4 @@
-#include "examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/tracker_api.cpp"
+#include "examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/tracker_api.cpp"
 
 #include <algorithm>
 #include <tuple>
