+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/utils/tracker_api.cpp
